refactor(signals): consolidate isTestFile/isCodeFile into gittensory-engine
isTestFile/isCodeFile were hand-ported three times (canonical in
src/signals/test-evidence.ts + path-matchers.ts, plus independent copies in
packages/gittensory-mcp/lib/local-branch.js and
packages/gittensory-mcp/scripts/gittensor-score-preview.mjs) -- commit
history has multiple "re-sync isTestFile with the server" fixes for exactly
this drift. A fourth, already-half-done copy was sitting unused in
packages/gittensory-engine/src/signals/test-evidence.ts (orphaned from
#3882, stale relative to the root file, never wired into the package's
exports or imported anywhere).

Makes the engine copy the actual up-to-date canonical source (adding the
extended isCodeFile alongside isTestPath/isSourcePath), and turns
src/signals/test-evidence.ts into a thin re-export shim -- same pattern
already used for scoring/preview.ts and focus-manifest.ts. path-matchers.ts
now delegates isCodeFile to the same source instead of composing its own
copy. Every existing call site keeps importing from the same paths.

The two packages/gittensory-mcp hand-ports aren't touched yet -- they can't
depend on @jsonbored/gittensory-engine until it's actually published to npm
(it currently only resolves via the workspace symlink), which needs a
one-time manual bootstrap outside CI. Follow-up once that's done.
